kafka/pyproject.toml
[tool.poetry]
name = "data_engineer_example_kafka"
version = "0.1.0"
description = ""
authors = ["mauza <mauza@pluralsight.com>"]

[tool.poetry.dependencies]
python = "^3.8"
confluent_kafka = "^1.5.0"
python-dotenv = "^0.14.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
